Moved deduplication logic into a helper class
